refactor: move testutils package to root
