+++ plugins/extractors/grafana/grafana_test.go
@@ -13,9 +13,9 @@ import (
 
 	"github.com/odpf/meteor/plugins"
 	"github.com/odpf/meteor/plugins/extractors/grafana"
-	"github.com/odpf/meteor/plugins/testutils"
 	"github.com/odpf/meteor/proto/odpf/assets"
 	"github.com/odpf/meteor/proto/odpf/assets/common"
+	"github.com/odpf/meteor/testutils"
 	"github.com/stretchr/testify/assert"
 )
